+++ docs/changes.rst
@@ -18,6 +18,9 @@ Unreleased.
  - Added :meth:`Image.percent_escape() <wand.image.BaseImage.percent_escape>` helper method. [:issue:`421`]
  - Added :meth:`Image.ping() <wand.image.Image.ping>` class method. [:issue:`#425`]
  - Added ``mean_color``, ``keep``, & ``remove`` parameters in :meth:`Image.connected_components() <wand.image.BaseImage.connected_components>` method.
+ - Added ``dissimilarity_threshold`` parameter, and renamed ``threshold`` parameter to ``similarity_threshold``
+   in :meth:`Image.similarity() <wand.image.BaseImage.similarity>` method.
+
 
 .. _changelog-0.5.5: